move counter handler and delete commented code
